Add files via upload
I added the scale to the graphs. 
Corrected the test for calculating delta before or after the obstacle. 
the linespplited needed to be casted to integer to be corrected tested
I added a function that generates a gnuplot script for the texture and deformation. 
So it is possible to edit the graphics without running the simulation from scratch.